+++ lib/Match.cppo.ml
@@ -688,7 +688,7 @@ let location_of_loc (loc : Location.t) source_path : Export.location =
       };
   }
 
-let search ~make_valid_source_path query_expr cmt =
+let search ~make_valid_source_path query_exprs cmt =
   (* We can't assume a single source file because a preprocessed file
      contains locations referring to more than one source file. *)
   let get_file_lines = memoize (Hashtbl.create 10) read_lines in
@@ -703,4 +703,4 @@ let search ~make_valid_source_path query_expr cmt =
         let e = max s (min num_lines loc_end.pos_lnum) in
         let lines = List.init (e - s + 1) (fun k -> src_lines.(s - 1 + k)) in
         Some { Export.location = location_of_loc loc source_path; lines })
-    (search_cmt query_expr cmt)
+    (List.concat_map (fun query_expr -> search_cmt query_expr cmt) query_exprs)